♻️ refactor: rename exercise-category files to kebab-case for consistency
- Rename exerciseCategory.* files to exercise-category.* in training module
- Update all import references across the codebase
- Maintain consistency with existing kebab-case naming convention
