+++ apps/api/src/modules/training/complex/complex.module.ts
@@ -2,7 +2,7 @@ import { MikroOrmModule } from '@mikro-orm/nestjs';
 import { Module, forwardRef } from '@nestjs/common';
 import { ComplexCategory } from '../complex-category/complex-category.entity';
 import { ComplexCategoryModule } from '../complex-category/complex-category.module';
-import { ExerciseCategoryModule } from '../exercise-category/exerciseCategory.module';
+import { ExerciseCategoryModule } from '../exercise-category/exercise-category.module';
 import { ExerciseComplex } from '../exercise-complex/exercise-complex.entity';
 import { Exercise } from '../exercise/exercise.entity';
 import { ExerciseModule } from '../exercise/exercise.module';
