+++ apps/api/src/modules/training/exercise-category/exercise-category.service.ts
