+++ apps/api/src/modules/training/exercise/exercise.module.ts
@@ -1,7 +1,7 @@
 import { MikroOrmModule } from '@mikro-orm/nestjs';
 import { Module } from '@nestjs/common';
 import { ExerciseCategory } from '../exercise-category/exercise-category.entity';
-import { ExerciseCategoryModule } from '../exercise-category/exerciseCategory.module';
+import { ExerciseCategoryModule } from '../exercise-category/exercise-category.module';
 import { ExerciseController } from './exercise.controller';
 import { Exercise } from './exercise.entity';
 import { ExerciseService } from './exercise.service';
