+++ apps/api/src/test/workout.e2e-spec.ts
@@ -16,7 +16,7 @@ import { AppModule } from '../app.module';
 import { createTestMikroOrmOptions } from '../config/mikro-orm.config';
 import { ComplexCategoryService } from '../modules/training/complex-category/complex-category.service';
 import { ComplexService } from '../modules/training/complex/complex.service';
-import { ExerciseCategoryService } from '../modules/training/exercise-category/exerciseCategory.service';
+import { ExerciseCategoryService } from '../modules/training/exercise-category/exercise-category.service';
 import { ExerciseService } from '../modules/training/exercise/exercise.service';
 import { WorkoutCategoryService } from '../modules/training/workout-category/workout-category.service';
 import { WORKOUT_ELEMENT_TYPES } from '../modules/training/workout-element/workout-element.entity';
